% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kwartielen.R
\name{kwartielen}
\alias{kwartielen}
\title{Eerste en derde kwartiel}
\usage{
kwartielen(x)
}
\arguments{
\item{x}{Een numerieke vector}
}
\value{
Een vector met het 25e en 75e percentiel
}
\description{
Eerste en derde kwartiel
}
