% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modus.R
\name{modus}
\alias{modus}
\title{Meest voorkomende waarde (modus)}
\usage{
modus(x)
}
\arguments{
\item{x}{Een vector}
}
\value{
De waarde die het meest voorkomt
}
\description{
Meest voorkomende waarde (modus)
}
