% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sam.R
\name{sam}
\alias{sam}
\title{Samenvattingstabel van verschillende statistische functies}
\usage{
sam(x)
}
\arguments{
\item{x}{Een object (vector, data frame, lijst, etc) waar de functies op toegepast worden.}
}
\value{
Een reactable tabel met functienaam, omschrijving en resultaat.
}
\description{
Samenvattingstabel van verschillende statistische functies
}
