% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samenvatting.R
\name{samenvatting}
\alias{samenvatting}
\title{Geef een samenvatting van de vector}
\usage{
samenvatting(x)
}
\arguments{
\item{x}{Een numerieke vector}
}
\value{
Een lijst met gemiddelde, mediaan en standaarddeviatie
}
\description{
Geef een samenvatting van de vector
}
