% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standaardviatie.R
\name{sd_berekenen}
\alias{sd_berekenen}
\title{Bereken de standaarddeviatie}
\usage{
sd_berekenen(x)
}
\arguments{
\item{x}{Een numerieke vector}
}
\value{
De standaarddeviatie
}
\description{
Bereken de standaarddeviatie
}
