% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sam.R
\name{show_functions_table}
\alias{show_functions_table}
\title{Toon overzicht van functies in het package}
\usage{
show_functions_table()
}
\value{
Een reactable-tabel met functienamen en argumenten.
}
\description{
Genereert een interactieve tabel met alle functies in het \code{ministate} package en hun argumenten.
}
\examples{
show_functions_table()
}
