% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samenvatting.R
\name{sam}
\alias{sam}
\title{Samenvattende statistieken van een vector}
\usage{
sam(x)
}
\arguments{
\item{x}{Een numerieke vector}
}
\value{
Een tibble met kernstatistieken
}
\description{
Samenvattende statistieken van een vector
}
