% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sam.R
\name{sam}
\alias{sam}
\title{Maak een interactieve samenvattingstabel van functie-uitkomsten}
\usage{
sam(resultaten, nrows = 5)
}
\arguments{
\item{resultaten}{Een lijst of data.frame met de uitkomsten van je functies.}

\item{nrows}{Aantal rijen per pagina (standaard 5).}
}
\value{
Een reactable widget met een overzichtelijke tabel.
}
\description{
Deze functie vat de resultaten van meerdere analyses samen in een overzichtelijke,
interactieve tabel met reactable.
}
