% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sam.R
\name{sam}
\alias{sam}
\title{Overzicht van functies in het package}
\usage{
sam()
}
\value{
Een reactable-tabel met functienamen en hun argumenten.
}
\description{
Genereert een interactieve tabel met alle functies in het \code{ministate} package en hun argumenten.
}
\examples{
sam()
}
