% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sam.R
\name{sam}
\alias{sam}
\title{Samenvattende statistiekentabel}
\usage{
sam(x)
}
\arguments{
\item{x}{Een numerieke vector}
}
\value{
Een reactable-tabel
}
\description{
Genereert een reactable-tabel met uitkomsten van alle statistiekfuncties op een vector.
}
\examples{
sam(c(1, 2, 2, 3, NA))
}
